Added a new route, created landing page
